+++ src/models/chats/chat.ts
@@ -33,6 +33,10 @@ import { ChatStore } from './chat-store';
 import Volume from '../volumes/volume';
 import File from '../files/file';
 import ReadReceipt from './read-receipt';
+
+// @ts-ignore to support desktop declarations emit until monorepo
+import Bluebird from 'bluebird';
+
 // to assign when sending a message and don't have an id yet
 let temporaryChatId = 0;
 function getTemporaryChatId() {
@@ -108,7 +112,7 @@ export default class Chat {
      */
     @observable.shallow limboMessages = [] as IObservableArray<Message>;
 
-    receipts: ObservableMap<ReadReceipt>;
+    receipts: ObservableMap<string, ReadReceipt>;
 
     // performance helper, to lookup messages by id and avoid duplicates
     _messageMap = {};
@@ -125,7 +129,8 @@ export default class Chat {
      */
     @computed
     get allParticipants(): IObservableArray<Contact> {
-        if (!this.db.boot || !this.db.boot.participants) return observable.shallowArray([]);
+        if (!this.db.boot || !this.db.boot.participants)
+            return observable.array([], { deep: false });
         return this.db.boot.participants.sort(this.compareContacts);
     }