+++ src/models/contacts/contact-store.ts
@@ -12,7 +12,7 @@ import Contact from './contact';
 import { setContactStore } from '../../helpers/di-contact-store';
 import MyContacts from '../contacts/my-contacts';
 import Invites from '../contacts/invites';
-import { EventEmitter } from 'eventemitter3';
+import EventEmitter from 'eventemitter3';
 import warnings from '../warnings';
 import createMap from '../../helpers/dynamic-array-map';
 import { getFirstLetterUpperCase } from './../../helpers/string';
@@ -194,7 +194,7 @@ export class ContactStore {
     });
 
     applyInvitesData = action(() => {
-        this.pendingContacts = observable.shallowArray(this.invites.issued);
+        this.pendingContacts = observable.array(this.invites.issued, { deep: false });
         when(
             () => this.invites.loaded && tofuStore.loaded && getChatStore().loaded,
             () => {