+++ CHANGELOG.md
@@ -2,6 +2,74 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="5.26.3"></a>
+## [5.26.3](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.26.2...v5.26.3) (2018-10-17)
+
+
+
+<a name="5.26.2"></a>
+## [5.26.2](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.26.1...v5.26.2) (2018-10-17)
+
+
+
+<a name="5.26.1"></a>
+## [5.26.1](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.26.0...v5.26.1) (2018-10-17)
+
+
+
+<a name="5.26.0"></a>
+# [5.26.0](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.8...v5.26.0) (2018-10-17)
+
+
+### Bug Fixes
+
+* **chat-invite-store:** async action ([6517a99](https://github.com/PeerioTechnologies/peerio-icebear/commit/6517a99))
+* cancel request and reconnect in setTimeout ([903510f](https://github.com/PeerioTechnologies/peerio-icebear/commit/903510f))
+* clear socket.io buffers ([1a03d7c](https://github.com/PeerioTechnologies/peerio-icebear/commit/1a03d7c))
+* imports ([1535440](https://github.com/PeerioTechnologies/peerio-icebear/commit/1535440))
+* make mixin not create enumerable properties on the prototype [ch14573] ([e4c8c03](https://github.com/PeerioTechnologies/peerio-icebear/commit/e4c8c03))
+* more deprecated mobx api ([94e8856](https://github.com/PeerioTechnologies/peerio-icebear/commit/94e8856))
+* tests ([00fea59](https://github.com/PeerioTechnologies/peerio-icebear/commit/00fea59))
+* tests should wait for user to be removed from boot keg ([ee72b4c](https://github.com/PeerioTechnologies/peerio-icebear/commit/ee72b4c))
+* upload progress ([40486ee](https://github.com/PeerioTechnologies/peerio-icebear/commit/40486ee))
+
+
+### Features
+
+* add disable onboarding emails flag to user settings ([9dbdfa6](https://github.com/PeerioTechnologies/peerio-icebear/commit/9dbdfa6))
+
+
+
+<a name="5.25.8"></a>
+## [5.25.8](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.7...v5.25.8) (2018-10-06)
+
+
+### Bug Fixes
+
+* don't export sendStored ([180524f](https://github.com/PeerioTechnologies/peerio-icebear/commit/180524f))
+* make User.current observable ([6e8cbce](https://github.com/PeerioTechnologies/peerio-icebear/commit/6e8cbce))
+* observe User.current in telemetry main ([476ff01](https://github.com/PeerioTechnologies/peerio-icebear/commit/476ff01))
+* review ([151c0fe](https://github.com/PeerioTechnologies/peerio-icebear/commit/151c0fe))
+* sendStored export ([941a851](https://github.com/PeerioTechnologies/peerio-icebear/commit/941a851))
+* use cross-fetch so telemetry works universally... ([67683c7](https://github.com/PeerioTechnologies/peerio-icebear/commit/67683c7))
+
+
+
+<a name="5.25.7"></a>
+## [5.25.7](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.6...v5.25.7) (2018-10-06)
+
+
+
+<a name="5.25.6"></a>
+## [5.25.6](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.5...v5.25.6) (2018-10-05)
+
+
+
+<a name="5.25.5"></a>
+## [5.25.5](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.4...v5.25.5) (2018-10-05)
+
+
+
 <a name="5.25.4"></a>
 ## [5.25.4](https://github.com/PeerioTechnologies/peerio-icebear/compare/v5.25.3...v5.25.4) (2018-09-28)